# remove precomiled sorts