TL;DR the cache of precompiled sort functions is rarely ever hit... would it be better to get rid of it?

The current implementation of `sort` dispatches to precompiled versions of `sort` when they use common comparison operators (e.g. `keyword<?`).

Perusing Racket code bases, it seems _extremely rare_ for someone to sort a list/vector using a _primitive comparison operator_ (e.g. string<?, keyword<?, etc) with no `#:key` argument.

Since providing a `#:key` argument or using a custom comparison function means we don't use one of the precompiled versions, it appears usages of `sort` _rarely_ benefit from any of the precompiled code.

In a quick experiment, I added a `printf` to `sort` and ran `raco setup` on a fresh install and it appears there were 40,311 cache misses and 33 cache hits (30 for keyword<?, 3 for string<? -- a hit rate of about 0.08%).

One last note, the size of the .zo file for `sort` is obviously impacted by this caching. Back in Racket v6.5 it was ~25k, recently after adding `vector-sort` and some additional precompiled versions the size was brought up to ~45k (i.e. the current size -- sorry, that's my fault!). Removing precompiled sorts altogether brings the .zo size down to ~15k.

Should we remove the precompiled versions (this PR does that)? Should we remove all but the most commonly (but still rather rarely) used comparators, e.g. `keyword<?`? Should we do nothing?