@rfindler 2017-10-18
Another option is to detect a commonly used comparator in the call site and
at that point construct the specialized version and deposit it into the
client module.

Also: check the shootout benchmarks.

And how much faster is the specialized version these days?