@pnwamk 2017-10-18
One note: after chatting w/ @samth I realized running `raco setup` may not necessarily reflect all common use cases of `sort` "in the wild" since it is just building all the modules, etc. So I'll have to do more sampling and some benchmarking, as @rfindler suggests to further clarify what might make sense.