@rfindler 2017-10-18
PS: I think it is great you're looking into this. This code isn't very recent and the compiler tradeoffs have changed since then (and are about to change again in a big way, of course!). I expect you'll find a point that makes a lot more sense nowadays.