@mfelleisen 2017-10-18
Would it make any sense to run all the benchmarks and tests we have with counting turned on? 

Would it make even more sense to equip drdr with ‘requests’ that collect bespoke statistics for the next N runs?