@pnwamk 2017-10-18
Shootout with Racket HEAD:
https://gist.github.com/pnwamk/4a4e34a9bf7f5f831def95f7984d300b

Shootout with this PR:
https://gist.github.com/pnwamk/e0acfafd6bdadf6198562d8c28c4b960

Only those shootout results that explicitly use `sort` in the source code:

```
moments (no obvious cache hits from source code)

with precompiled sorts
----------------------
[racket moments (2288 2285 320) 12]
[racket moments (2292 2288 320) 12]
[racket moments (2308 2304 328) 12]

without precompiled sorts
-------------------------
[racket moments (2316 2312 332) 15]
[racket moments (2284 2284 316) 15]
[racket moments (2292 2292 324) 15]






k-nucleotide (no obvious cache hits from source code)

with precompiled sorts
----------------------
[racket k-nucleotide (2936 2936 56) 18]
[racket k-nucleotide (2940 2937 52) 18]
[racket k-nucleotide (2916 2910 52) 18]


without precompiled sorts
-------------------------
[racket k-nucleotide (2924 2921 52) 16]
[racket k-nucleotide (2940 2938 52) 16]
[racket k-nucleotide (2916 2913 48) 16]






meteor (should hit a cached version of sort for `>`)

with precompiled sorts
----------------------
[racket meteor (296 298 0) 47]
[racket meteor (300 298 0) 47]
[racket meteor (300 300 0) 47]


without precompiled sorts
-------------------------
[racket meteor (300 299 0) 49]
[racket meteor (304 304 0) 49]
[racket meteor (296 299 0) 49]



wordfreq (should hit a cached sort for `string>?`)

with precompiled sorts
----------------------
[racket wordfreq (2748 2749 20) 69]
[racket wordfreq (2812 2809 16) 69]
[racket wordfreq (2744 2741 16) 69]


without precompiled sorts
-------------------------
[racket wordfreq (2824 2825 20) 68]
[racket wordfreq (2856 2856 24) 68]
[racket wordfreq (2808 2806 20) 68]
```