@pnwamk 2017-10-18
a quickly thrown together microbenchmark of just sorting with a comparison operator that has a precompiled version in the current `sort` implementation:

```racket
#lang racket/base


(define short-lists '(()
                      (1)
                      (1 2)
                      (2 1)
                      (1 2 3)
                      (3 2 1)
                      (1 2 3 4)
                      (5 4 3 2 1)
                      (1 2 3 4 5 6 7 8 9 10)
                      (10 9 8 7 6 5 4 3 2 1)))

(define list100 (build-list 100 (λ _ (random 100))))
(define list1000 (build-list 1000 (λ _ (random 1000))))
(define list10000 (build-list 10000 (λ _ (random 10000))))


(printf "sorting short lists:\n")
(time (for* ([_ (in-range 100000)]
             [l (in-list short-lists)])
        (sort l <)))

(printf "sorting list of length 100:\n")
(time (for* ([_ (in-range 10000)])
        (sort list100 <)))

(printf "sorting list of length 1000:\n")
(time (for* ([_ (in-range 10000)])
        (sort list1000 <)))

(printf "sorting list of length 10000:\n")
(time (for* ([_ (in-range 10000)])
        (sort list10000 <)))
```

results with precompiled sorts:

```
> racket sorting.rkt      
sorting short lists:
cpu time: 140 real time: 140 gc time: 4
sorting list of length 100:
cpu time: 56 real time: 55 gc time: 0
sorting list of length 1000:
cpu time: 1100 real time: 1100 gc time: 4
sorting list of length 10000:
cpu time: 15148 real time: 15145 gc time: 52
```


results without precompiled sorts:

```
> racket sorting.rkt      
sorting short lists:
cpu time: 164 real time: 165 gc time: 0
sorting list of length 100:
cpu time: 128 real time: 126 gc time: 0
sorting list of length 1000:
cpu time: 1812 real time: 1811 gc time: 20
sorting list of length 10000:
cpu time: 23948 real time: 23940 gc time: 52
```