@pnwamk 2017-10-18
Obviously in the microbenchmarks, replacing `<` with `(λ (x y) (< x y))` causes both to run in the same time (since the comparison function has no precompiled version).

Also, obviously the microbenchmarks ignore the cache hit rate (and I still don't have numbers on what that looks like outside of just running `raco make`, but I worry it is common to use a custom comparison function or a #:key argument -- but I could be wrong).