@mflatt 2017-10-18
I offer no conclusion, but FWIW: These precompiled sorts were a hassle for the linklet flattener, so I considered removing them at the time. I decided to keep them after seeing the difference in microbenchmarks, but only because that seemed safer. If no real code hits the cache, that's a good argument for getting rid of it.