@pnwamk 2017-10-19
I threw together a quick and dirty script that checks for usages of `(sort ...)` and tries to determine if they would be a cache hit or miss based on whether or not keyword args were present, and what the literal symbol for the sort function was:

https://gist.github.com/pnwamk/5f507da57a3276da81f58eb3a24e170f

Here are the results from running that on all the .rkt files (18,058) from pkgs.racket-lang.org:

```
total usages of sort found: 915
confirmed cache misses (i.e. used #:key): 342
probable cache misses (i.e. symbol didn't match precompile list): 317
cache hits (symbol matched a precompiled function): 234
unknown (i.e. didn't match syntax for normal sort usage): 22
hit counts for precompiled versions:
((char<? . 3)
 (< . 96)
 (string<? . 72)
 (<= . 2)
 (string>? . 6)
 (>= . 2)
 (keyword<? . 14)
 (string-ci<=? . 3)
 (string<=? . 22)
 (> . 13)
 (string-ci<? . 1))
```

Again, this doesn't really indicate how _often_ the cache is being hit or missed during execution, and it doesn't exclude usages that are uninteresting (i.e. some are trivial tests that merely sort a list of symbols so the test can compare two lists w/o worrying about order, etc).