@bennn 2017-10-27
@pnwamk I think you should email the racket-dev list about this pull request. (Maybe someone there has strong options / requirements, but hasn't seen this yet.)